pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "story-scraper"
version = "0.1.0"
description = "Web scraper and labeling tool for story sites (Selenium)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
story-scraper = "story_scraper.scraper:main"
story-labeler = "story_scraper.labeler:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["story_scraper*"]
